+++ src/main/java/com/lbogdanandrei/cardealerapp/model/DealerModel.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import org.springframework.lang.NonNull;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -14,7 +13,7 @@ public class DealerModel {
 
     @Id
     @GeneratedValue(strategy =  GenerationType.IDENTITY)
-    private BigInteger id;
+    private int id;
 
     @Column
     @NonNull
